Update Gateway API to v1.0.0 (#1250)
Problems:
- NGF does not support v1.0.0 of Gateway API.
- NGF does not use the official PartiallyInvalid 
condition from Gateway API SDK.
- The Gateway API validating webhook should 
only be installed when Kubernetes version is <1.25

Solutions:
- Update to v1.0.0 of Gateway API
- Replace custom PartiallyInvalid condition with 
the Gateway API condition.
- Add logic to pipeline and conformance scripts 
to conditionally install the webhook.

+++ conformance/scripts/uninstall-gateway.sh
@@ -9,9 +9,14 @@ if [ $1 == "main" ]; then
     temp_dir=$(mktemp -d)
     cd ${temp_dir}
     curl -s https://codeload.github.com/kubernetes-sigs/gateway-api/tar.gz/main | tar -xz --strip=2 gateway-api-main/config
-    kubectl delete -f webhook
     kubectl delete -f crd/standard
+    if [ $2 == "true" ]; then
+        kubectl delete -f webhook
+    fi
     rm -rf ${temp_dir}
 else
     kubectl delete -f https://github.com/kubernetes-sigs/gateway-api/releases/download/v$1/standard-install.yaml
+    if [ $2 == "true" ]; then
+        kubectl delete -f https://github.com/kubernetes-sigs/gateway-api/releases/download/v$1/webhook-install.yaml
+    fi
 fi
